capstone/requirements.txt
Django==3.2.3
Pillow==8.2.0
channels==3.0.3
daphne==3.0.2
sqlparse==0.4.1
twilio==6.59.0
django-environ==0.4.5
selenium
django-import-export
